установил значения по умолчанию
