+++ Industrial/Industrial/Profile/UserService.swift
@@ -1,0 +1,16 @@
+//
+//  UserService.swift
+//  Industrial
+//
+//  Created by DmitriiG on 04.08.2022.
+//
+
+import Foundation
+import UIKit
+
+protocol UserService {
+    
+    func userInfo(name: String) -> User?
+    
+}
+
